requirement.txt
fastapi
uvicorn[standard]
transformers>=4.40.0.dev0
git+https://github.com/UKPLab/sentence-transformers.git
qdrant-client
langgraph      
Pillow
python-multipart
requests
PyYAML
addict
matplotlib


PyPDF2 
pdf2image 
pytesseract 
easyocr  
pyyaml
bitsandbytes
accelerate